# control-plane/internal/apispec/openapi.yaml
openapi: 3.1.0
info:
  title: Hive API
  version: 0.5.0
  description: 'Hive is a Swarm-native platform control plane. This spec is the single

    source of truth for all API contracts. Every request body, response, and

    error shape is defined here.

    '
servers:
- url: /api/v1
  description: Control plane API v1
components:
  schemas:
    ErrorResponse:
      description: ErrorResponse schema.
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
      example:
        code: bad_request
        message: invalid payload
    StatusResponse:
      description: StatusResponse schema.
      type: object
      additionalProperties: false
      required:
      - status
      properties:
        status:
          type: string
          description: Current status.
      example:
        status: ok
...
    WebhookPayload:
      type: object
      additionalProperties: true
      description: Provider-specific webhook payload (GitHub, GitLab, Bitbucket, Gitea).
...
    PreviewDeployment:
      description: Preview deployment for a pull request
      type: object
      required:
      - id
      - applicationId
      - prNumber
      - branch
      - status
      - url
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Preview deployment ID
        applicationId:
          type: string
          format: uuid
          description: Application ID
        prNumber:
          type: integer
          description: Pull request number
        branch:
          type: string
          description: Git branch name
        commitSha:
          type: string
          description: Git commit SHA
        status:
          type: string
          enum:
          - building
          - deploying
          - ready
          - failed
          - expired
          description: Preview status
        url:
          type: string
          description: Preview URL
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        applicationId: 550e8400-e29b-41d4-a716-446655440001
        prNumber: 42
        branch: feature/new-ui
        commitSha: abc123
        status: ready
        url: https://pr-42--myapp.preview.example.com
        expiresAt: '2026-01-08T00:00:00Z'
        createdAt: '2026-01-01T00:00:00Z'
    PreviewDeploymentList:
      description: List of preview deployments
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PreviewDeployment'
    PreviewDeploymentCreateRequest:
      description: Request to create a preview deployment
      type: object
      required:
      - prNumber
      - branch
      properties:
        prNumber:
          type: integer
          description: Pull request number
        branch:
          type: string
          description: Git branch name
        commitSha:
          type: string
          description: Git commit SHA
      example:
        prNumber: 42
        branch: feature/new-ui
        commitSha: abc123
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      description: bearerAuth schema.
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /webhooks/github:
    post:
      operationId: githubWebhook
      summary: GitHub webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '202':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
      description: Create webhooks github
      security: []
  /webhooks/gitlab:
    post:
      operationId: gitlabWebhook
      summary: GitLab webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '202':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
      description: Create webhooks gitlab
      security: []
  /webhooks/bitbucket:
    post:
      operationId: bitbucketWebhook
      summary: Bitbucket webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '202':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
      description: Create webhooks bitbucket
      security: []
  /webhooks/gitea:
    post:
      operationId: giteaWebhook
      summary: Gitea webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '202':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
      description: Create webhooks gitea
      security: []
...
  /applications/{id}/previews:
    get:
      operationId: listPreviewDeployments
      summary: List preview deployments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewDeploymentList'
      description: List preview deployments for an application
      security:
      - bearerAuth: []
    post:
      operationId: createPreviewDeployment
      summary: Create a preview deployment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewDeploymentCreateRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required:
                - id
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
      description: Create a preview deployment for a pull request
      security:
      - bearerAuth: []
  /applications/{id}/previews/{previewId}:
    get:
      operationId: getPreviewDeployment
      summary: Get preview deployment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: previewId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewDeployment'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Get preview deployment by ID
      security:
      - bearerAuth: []
    delete:
      operationId: deletePreviewDeployment
      summary: Delete preview deployment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: previewId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                description: No response body.
        '404':
          $ref: '#/components/responses/NotFound'
      description: Delete a preview deployment
      security:
      - bearerAuth: []
...